@@ -8,6 +8,7 @@ from distutils import log
 from types import CodeType
 import sys
 import os
+import re
 import textwrap
 import marshal
 
@@ -37,6 +38,16 @@ def strip_module(filename):
     if filename.endswith('module'):
         filename = filename[:-6]
     return filename
+
+
+def sorted_walk(dir):
+    """Do os.walk in a reproducible way,
+    independent of indeterministic filesystem readdir order
+    """
+    for base, dirs, files in os.walk(dir):
+        dirs.sort()
+        files.sort()
+        yield base, dirs, files
 
 
 def write_stub(resource, pyfile):
@@ -232,10 +243,27 @@ class bdist_egg(Command):
         log.info("Removing .py files from temporary directory")
         for base, dirs, files in walk_egg(self.bdist_dir):
             for name in files:
+                path = os.path.join(base, name)
+
                 if name.endswith('.py'):
-                    path = os.path.join(base, name)
                     log.debug("Deleting %s", path)
                     os.unlink(path)
+
+                if base.endswith('__pycache__'):
+                    path_old = path
+
+                    pattern = r'(?P<name>.+)\.(?P<magic>[^.]+)\.pyc'
+                    m = re.match(pattern, name)
+                    path_new = os.path.join(
+                        base, os.pardir, m.group('name') + '.pyc')
+                    log.info(
+                        "Renaming file from [%s] to [%s]"
+                        % (path_old, path_new))
+                    try:
+                        os.remove(path_new)
+                    except OSError:
+                        pass
+                    os.rename(path_old, path_new)
 
     def zip_safe(self):
         safe = getattr(self.distribution, 'zip_safe', None)
@@ -302,7 +330,7 @@ class bdist_egg(Command):
         ext_outputs = []
 
         paths = {self.bdist_dir: ''}
-        for base, dirs, files in os.walk(self.bdist_dir):
+        for base, dirs, files in sorted_walk(self.bdist_dir):
             for filename in files:
                 if os.path.splitext(filename)[1].lower() in NATIVE_EXTENSIONS:
                     all_outputs.append(paths[base] + filename)
@@ -329,7 +357,7 @@ NATIVE_EXTENSIONS = dict.fromkeys('.dll .so .dylib .pyd'.split())
 
 def walk_egg(egg_dir):
     """Walk an unpacked egg's contents, skipping the metadata directory"""
-    walker = os.walk(egg_dir)
+    walker = sorted_walk(egg_dir)
     base, dirs, files = next(walker)
     if 'EGG-INFO' in dirs:
         dirs.remove('EGG-INFO')
@@ -385,8 +413,10 @@ def scan_module(egg_dir, base, name, stubs):
     module = pkg + (pkg and '.' or '') + os.path.splitext(name)[0]
     if sys.version_info < (3, 3):
         skip = 8  # skip magic & date
-    else:
+    elif sys.version_info < (3, 7):
         skip = 12  # skip magic & date & file size
+    else:
+        skip = 16  # skip magic & reserved? & date & file size
     f = open(filename, 'rb')
     f.read(skip)
     code = marshal.load(f)
@@ -463,10 +493,10 @@ def make_zipfile(zip_filename, base_dir, verbose=0, dry_run=0, compress=True,
     compression = zipfile.ZIP_DEFLATED if compress else zipfile.ZIP_STORED
     if not dry_run:
         z = zipfile.ZipFile(zip_filename, mode, compression=compression)
-        for dirname, dirs, files in os.walk(base_dir):
+        for dirname, dirs, files in sorted_walk(base_dir):
             visit(z, dirname, files)
         z.close()
     else:
-        for dirname, dirs, files in os.walk(base_dir):
+        for dirname, dirs, files in sorted_walk(base_dir):
             visit(None, dirname, files)
     return zip_filename
