@@ -95,7 +95,9 @@ class develop(namespaces.DevelopInstaller, easy_install):
         path_to_setup = egg_base.replace(os.sep, '/').rstrip('/')
         if path_to_setup != os.curdir:
             path_to_setup = '../' * (path_to_setup.count('/') + 1)
-        resolved = normalize_path(os.path.join(install_dir, egg_path, path_to_setup))
+        resolved = normalize_path(
+            os.path.join(install_dir, egg_path, path_to_setup)
+        )
         if resolved != normalize_path(os.curdir):
             raise DistutilsOptionError(
                 "Can't get a consistent path to setup script from"
