@@ -8,6 +8,7 @@ import site
 import sys
 import io
 import subprocess
+import platform
 
 from setuptools.extern import six
 from setuptools.command import test
@@ -61,7 +62,8 @@ class TestDevelop:
     in_virtualenv = hasattr(sys, 'real_prefix')
     in_venv = hasattr(sys, 'base_prefix') and sys.base_prefix != sys.prefix
 
-    @pytest.mark.skipif(in_virtualenv or in_venv,
+    @pytest.mark.skipif(
+        in_virtualenv or in_venv,
         reason="Cannot run when invoked in a virtualenv or venv")
     def test_2to3_user_mode(self, test_env):
         settings = dict(
@@ -101,7 +103,8 @@ class TestDevelop:
         Test that console scripts are installed and that they reference
         only the project by name and not the current version.
         """
-        pytest.skip("TODO: needs a fixture to cause 'develop' "
+        pytest.skip(
+            "TODO: needs a fixture to cause 'develop' "
             "to be invoked without mutating environment.")
         settings = dict(
             name='foo',
@@ -153,8 +156,14 @@ class TestNamespaces:
             with test.test.paths_on_pythonpath([str(target)]):
                 subprocess.check_call(develop_cmd)
 
-    @pytest.mark.skipif(bool(os.environ.get("APPVEYOR")),
-        reason="https://github.com/pypa/setuptools/issues/851")
+    @pytest.mark.skipif(
+        bool(os.environ.get("APPVEYOR")),
+        reason="https://github.com/pypa/setuptools/issues/851",
+    )
+    @pytest.mark.skipif(
+        platform.python_implementation() == 'PyPy' and six.PY3,
+        reason="https://github.com/pypa/setuptools/issues/1202",
+    )
     def test_namespace_package_importable(self, tmpdir):
         """
         Installing two packages sharing the same namespace, one installed
@@ -167,6 +176,8 @@ class TestNamespaces:
         target = tmpdir / 'packages'
         # use pip to install to the target directory
         install_cmd = [
+            sys.executable,
+            '-m',
             'pip',
             'install',
             str(pkg_A),
